svc_encoder_rtc.cc: fix conversion warning
w/various versions of Visual Studio, after:
5665f1cd6d svc_encoder_rtc: improve multilayer metadata handling

Rightsize mantissa_len to `uint8_t`, it's documented as only containing
values between 1 and 32.

fixes:
examples\svc_encoder_rtc.cc(1416,62): warning C4244:
  'argument': conversion from 'const uint16_t' to 'uint8_t', possible
  loss of data

+++ examples/multilayer_metadata.cc
@@ -896,7 +896,7 @@ bool double_to_depth_representation_element(
   assert(v >= 1.0 && v < 2.0);
   v -= 1.0;
   uint32_t mantissa = 0;
-  uint16_t mantissa_len = 0;
+  uint8_t mantissa_len = 0;
   constexpr uint16_t kMaxMantissaLen = 32;
   do {
     const int bit = (v >= 0.5);
